# mama-app-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mama na Mwana API Documentation",
        "description": "API documentation for Mama na Mwana application - A comprehensive maternal health tracking system",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/mama-data": {
            "post": {
                "tags": [
                    "Mama Data"
                ],
                "summary": "Store new mama's pregnancy data",
                "description": "Creates a new pregnancy record for the mama user",
                "operationId": "storeMamaData",
                "requestBody": {
                    "description": "Mama's pregnancy information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "age_group",
                                    "due_date",
                                    "gestational_period"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the user",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "first_child": {
                                        "description": "Whether this is the first child",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "age_group": {
                                        "description": "Mother's age group",
                                        "type": "string",
                                        "enum": [
                                            "18-24 years old",
                                            "25-34 years old",
                                            "35-44 years old",
                                            "44 years old or above"
                                        ],
                                        "example": "25-34 years old"
                                    },
                                    "due_date": {
                                        "description": "Expected delivery date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-31"
                                    },
                                    "gestational_period": {
                                        "description": "Current pregnancy week (1-42 weeks)",
                                        "type": "integer",
                                        "maximum": 42,
                                        "minimum": 1,
                                        "example": 28
                                    },
                                    "baby_gender": {
                                        "description": "Known or expected gender of the baby",
                                        "type": "string",
                                        "enum": [
                                            "boy",
                                            "girl",
                                            "i don't know yet"
                                        ],
                                        "example": "i don't know yet"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pregnancy data successfully recorded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mama data created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MamaData"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MamaData": {
                "required": [
                    "user_id",
                    "age_group",
                    "due_date",
                    "gestational_period"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the mama data record",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "description": "The ID of the user this data belongs to",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "first_child": {
                        "description": "Whether this is the first child",
                        "type": "boolean",
                        "example": true
                    },
                    "age_group": {
                        "description": "Mother's age group",
                        "type": "string",
                        "enum": [
                            "18-24 years old",
                            "25-34 years old",
                            "35-44 years old",
                            "44 years old or above"
                        ],
                        "example": "25-34 years old"
                    },
                    "due_date": {
                        "description": "Expected delivery date",
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    },
                    "gestational_period": {
                        "description": "Current pregnancy week (1-42 weeks)",
                        "type": "integer",
                        "maximum": 42,
                        "minimum": 1,
                        "example": 28
                    },
                    "baby_gender": {
                        "description": "Known or expected gender of the baby",
                        "type": "string",
                        "enum": [
                            "boy",
                            "girl",
                            "i don't know yet"
                        ],
                        "example": "i don't know yet"
                    },
                    "created_at": {
                        "description": "Timestamp when the record was created",
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-03-13T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the record was last updated",
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-03-13T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email/phone and password to get the authentication token",
                "name": "Token Based Authentication",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}